Additional tidying up
